+++ CHANGELOG.md
@@ -3,11 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- breaking: use pretty URLs (`$f.md` converts to `$f/index.html` instead of `$f.html`)
+- breaking: rename `sitedown.fileToPageBody` to `sitedown.mdToHtml`
+- breaking: use single `layout.html` instead of `header.html` and `footer.html`
+  - parsed markdown content is injected into `.markdown-body`
+  - text from first `h1` is injected into `title`
+
 ## [1.1.0]
-* avoid rewriting external links to MDs ([#3](https://github.com/ngoldman/sitedown/pull/3))
-* generate CSS classes for highlightjs support ([#4](https://github.com/ngoldman/sitedown/pull/4))
+- avoid rewriting external links to MDs ([#3](https://github.com/ngoldman/sitedown/pull/3))
+- generate CSS classes for highlightjs support ([#4](https://github.com/ngoldman/sitedown/pull/4))
 
 ## 1.0.0
-* engage
+- engage
 
 [1.1.0]: https://github.com/ngoldman/sitedown/compare/v1.0.0...v1.1.1